+++ node/src/p2p/mod.rs
@@ -111,5 +111,6 @@ impl_into_global_action!(p2p::identify::P2pIdentifyAction);
 impl_into_global_action!(p2p::P2pNetworkSelectAction);
 impl_into_global_action!(p2p::P2pNetworkPnetAction);
 impl_into_global_action!(p2p::P2pNetworkNoiseAction);
+impl_into_global_action!(p2p::P2pNetworkRpcAction);
 
 impl p2p::P2pActionTrait<crate::State> for crate::Action {}
