+++ p2p/src/network/rpc/mod.rs
@@ -6,6 +6,3 @@ pub use self::p2p_network_rpc_state::{P2pNetworkRpcState, RpcMessage};
 
 #[cfg(feature = "p2p-libp2p")]
 mod p2p_network_rpc_reducer;
-
-#[cfg(feature = "p2p-libp2p")]
-mod p2p_network_rpc_effects;
