+++ p2p/src/network/rpc/p2p_network_rpc_reducer.rs
@@ -1,33 +1,83 @@
-use binprot::BinProtRead;
-use mina_p2p_messages::rpc_kernel::{MessageHeader, QueryHeader, RpcMethod};
+use std::sync::Arc;
 
-use crate::{Limit, P2pLimits};
+use binprot::BinProtRead;
+use mina_p2p_messages::{
+    rpc,
+    rpc_kernel::{
+        MessageHeader, PayloadBinprotReader as _, QueryHeader, ResponseHeader, RpcMethod,
+        RpcQueryReadError, RpcResponseReadError,
+    },
+    v2,
+    versioned::Ver,
+};
+use openmina_core::{error, fuzz_maybe, fuzzed_maybe, Substate};
+use redux::Dispatcher;
+
+use crate::{
+    channels::rpc::{
+        BestTipWithProof, P2pChannelsRpcAction, P2pRpcRequest, P2pRpcResponse,
+        StagedLedgerAuxAndPendingCoinbases,
+    },
+    connection::outgoing::P2pConnectionOutgoingInitOpts,
+    disconnection::{P2pDisconnectionAction, P2pDisconnectionReason},
+    Data, Limit, P2pLimits, P2pNetworkState, P2pNetworkYamuxAction, PeerId,
+};
 
 use self::p2p_network_rpc_state::P2pNetworkRpcError;
 
 use super::*;
 
 impl P2pNetworkRpcState {
-    pub fn reducer(
-        &mut self,
+    pub fn reducer<State, Action>(
+        mut state_context: Substate<Action, State, P2pNetworkState>,
         action: redux::ActionWithMeta<&P2pNetworkRpcAction>,
         limits: &P2pLimits,
-    ) {
-        match action.action() {
-            P2pNetworkRpcAction::Init { incoming, .. } => {
-                self.is_incoming = *incoming;
+    ) -> Result<(), String>
+    where
+        State: crate::P2pStateTrait,
+        Action: crate::P2pActionTrait<State>,
+    {
+        let rpc_state = state_context
+            .get_substate_mut()?
+            .find_rpc_state_mut(action.action())
+            .ok_or_else(|| format!("RPC state not found for action: {:?}", action.action()))?;
+
+        let (action, meta) = action.split();
+        match action {
+            P2pNetworkRpcAction::Init {
+                incoming,
+                addr,
+                peer_id,
+                stream_id,
+            } => {
+                rpc_state.is_incoming = *incoming;
+
+                let dispatcher = state_context.into_dispatcher();
+                dispatcher.push(P2pNetworkRpcAction::OutgoingData {
+                    addr: *addr,
+                    peer_id: *peer_id,
+                    stream_id: *stream_id,
+                    data: Data::from(RpcMessage::Handshake.into_bytes()),
+                    fin: false,
+                });
+                Ok(())
             }
-            P2pNetworkRpcAction::IncomingData { data, .. } => {
-                self.buffer.extend_from_slice(data);
+            P2pNetworkRpcAction::IncomingData {
+                data,
+                addr,
+                peer_id,
+                stream_id,
+            } => {
+                rpc_state.buffer.extend_from_slice(data);
                 let mut offset = 0;
                 // TODO(akoptelov): there shouldn't be the case where we have multiple incoming messages at once (or at least other than heartbeat)
                 loop {
-                    let buf = &self.buffer[offset..];
+                    let buf = &rpc_state.buffer[offset..];
                     if let Some(len_bytes) = buf.get(..8).and_then(|s| s.try_into().ok()) {
                         let len = u64::from_le_bytes(len_bytes) as usize;
-                        if let Err(err) = self.check_rpc_limit(len, limits) {
-                            self.error = Some(err);
-                            return;
+                        if let Err(err) = rpc_state.check_rpc_limit(len, limits) {
+                            rpc_state.error = Some(err);
+                            break;
                         }
                         if buf.len() >= 8 + len {
                             offset += 8 + len;
@@ -48,53 +98,233 @@ impl P2pNetworkRpcState {
                                     bytes: slice.to_vec().into(),
                                 },
                                 Err(err) => {
-                                    self.error = Some(P2pNetworkRpcError::Binprot(err.to_string()));
+                                    rpc_state.error =
+                                        Some(P2pNetworkRpcError::Binprot(err.to_string()));
                                     continue;
                                 }
                             };
-                            self.incoming.push_back(msg);
+                            rpc_state.incoming.push_back(msg);
                             continue;
                         }
                     }
 
+                    if offset != 0 {
+                        rpc_state.buffer = rpc_state.buffer[offset..].to_vec();
+                    }
                     break;
                 }
 
-                if offset != 0 {
-                    self.buffer = self.buffer[offset..].to_vec();
+                let incoming = rpc_state.incoming.front().cloned();
+                let dispatcher = state_context.into_dispatcher();
+
+                if let Some(message) = incoming {
+                    dispatcher.push(P2pNetworkRpcAction::IncomingMessage {
+                        addr: *addr,
+                        peer_id: *peer_id,
+                        stream_id: *stream_id,
+                        message,
+                    })
                 }
+
+                Ok(())
             }
-            P2pNetworkRpcAction::IncomingMessage { message, .. } => {
+            action @ P2pNetworkRpcAction::IncomingMessage {
+                message,
+                addr,
+                peer_id,
+                stream_id,
+            } => {
                 if let RpcMessage::Response { header, .. } = message {
-                    if let Some(QueryHeader { id, tag, version }) = &self.pending {
-                        *self.total_stats.entry((tag.clone(), *version)).or_default() += 1;
+                    if let Some(QueryHeader { id, tag, version }) = &rpc_state.pending {
+                        *rpc_state
+                            .total_stats
+                            .entry((tag.clone(), *version))
+                            .or_default() += 1;
                         if id != &header.id {
-                            openmina_core::error!(action.time(); "receiving response with wrong id: {}", header.id);
+                            openmina_core::error!(meta.time(); "receiving response with wrong id: {}", header.id);
                         }
                     } else {
-                        openmina_core::error!(action.time(); "receiving response without query");
+                        openmina_core::error!(meta.time(); "receiving response without query");
                     }
                 } else if let RpcMessage::Query { header, .. } = message {
-                    if self.pending.is_none() {
-                        self.pending = Some(header.clone());
+                    if rpc_state.pending.is_none() {
+                        rpc_state.pending = Some(header.clone());
                     } else {
-                        openmina_core::error!(action.time(); "receiving query while another query is pending");
+                        openmina_core::error!(meta.time(); "receiving query while another query is pending");
                     }
                 }
 
-                self.incoming.pop_front();
+                rpc_state.incoming.pop_front();
+
+                let (dispatcher, state) = state_context.into_dispatcher_and_state();
+                let network_state: &P2pNetworkState = state.substate()?;
+                let state = network_state
+                    .find_rpc_state(action)
+                    .ok_or_else(|| format!("RPC state not found for action: {:?}", action))?;
+
+                match &message {
+                    RpcMessage::Handshake => {
+                        if !state.is_incoming {
+                            dispatcher.push(P2pChannelsRpcAction::Ready { peer_id: *peer_id });
+                        }
+                    }
+                    RpcMessage::Heartbeat => {}
+                    RpcMessage::Query { header, bytes } => {
+                        if let Err(e) = rpc_query_effects(*peer_id, header, bytes, dispatcher) {
+                            dispatcher.push(P2pDisconnectionAction::Init {
+                                peer_id: *peer_id,
+                                reason: P2pDisconnectionReason::P2pChannelReceiveFailed(
+                                    e.to_string(),
+                                ),
+                            });
+                        }
+                    }
+                    RpcMessage::Response {
+                        header: ResponseHeader { id },
+                        bytes,
+                    } => {
+                        let query_header = match state.pending.as_ref() {
+                            Some(header) if &header.id == id => header.clone(),
+                            Some(header) => {
+                                error!(meta.time(); "received response with it {} while expecting {id}", header.id);
+                                return Ok(());
+                            }
+                            None => {
+                                error!(meta.time(); "received response while no query is sent");
+                                return Ok(());
+                            }
+                        };
+                        // unset pending
+                        dispatcher.push(P2pNetworkRpcAction::PrunePending {
+                            peer_id: *peer_id,
+                            stream_id: *stream_id,
+                        });
+
+                        if let Err(e) =
+                            rpc_response_effects(*peer_id, &query_header, bytes, dispatcher)
+                        {
+                            dispatcher.push(P2pDisconnectionAction::Init {
+                                peer_id: *peer_id,
+                                reason: P2pDisconnectionReason::P2pChannelReceiveFailed(
+                                    e.to_string(),
+                                ),
+                            });
+                        }
+                    }
+                }
+
+                if let Some(message) = state.incoming.front().cloned() {
+                    dispatcher.push(P2pNetworkRpcAction::IncomingMessage {
+                        addr: *addr,
+                        peer_id: *peer_id,
+                        stream_id: *stream_id,
+                        message,
+                    });
+                }
+                Ok(())
             }
             P2pNetworkRpcAction::PrunePending { .. } => {
-                self.pending = None;
+                rpc_state.pending = None;
+                Ok(())
             }
-            P2pNetworkRpcAction::HeartbeatSend { .. } => {
-                self.last_heartbeat_sent = Some(action.time());
+            P2pNetworkRpcAction::HeartbeatSend {
+                addr,
+                peer_id,
+                stream_id,
+            } => {
+                rpc_state.last_heartbeat_sent = Some(meta.time());
+
+                let dispatcher = state_context.into_dispatcher();
+
+                dispatcher.push(P2pNetworkRpcAction::OutgoingData {
+                    addr: *addr,
+                    peer_id: *peer_id,
+                    stream_id: *stream_id,
+                    data: Data::from(RpcMessage::Heartbeat.into_bytes()),
+                    fin: false,
+                });
+
+                Ok(())
             }
-            P2pNetworkRpcAction::OutgoingQuery { query, .. } => {
-                self.last_id = query.id;
-                self.pending = Some(query.clone());
+            P2pNetworkRpcAction::OutgoingQuery {
+                query,
+                data,
+                peer_id,
+            } => {
+                rpc_state.last_id = query.id;
+                rpc_state.pending = Some(query.clone());
+
+                let addr = rpc_state.addr;
+                let stream_id = rpc_state.stream_id;
+                let dispatcher = state_context.into_dispatcher();
+                dispatcher.push(P2pNetworkRpcAction::OutgoingData {
+                    addr,
+                    peer_id: *peer_id,
+                    stream_id,
+                    data: Data::from(
+                        RpcMessage::Query {
+                            header: query.clone(),
+                            bytes: data.clone(),
+                        }
+                        .into_bytes(),
+                    ),
+                    fin: false,
+                });
+
+                Ok(())
             }
-            _ => {}
+            P2pNetworkRpcAction::OutgoingData {
+                addr,
+                stream_id,
+                data,
+                ..
+            } => {
+                let dispatcher = state_context.into_dispatcher();
+                let mut data = data.clone();
+                fuzz_maybe!(&mut data, crate::fuzzer::mutate_rpc_data);
+                let flags = fuzzed_maybe!(Default::default(), crate::fuzzer::mutate_yamux_flags);
+
+                dispatcher.push(P2pNetworkYamuxAction::OutgoingData {
+                    addr: *addr,
+                    stream_id: *stream_id,
+                    data,
+                    flags,
+                });
+
+                Ok(())
+            }
+            P2pNetworkRpcAction::OutgoingResponse {
+                peer_id,
+                response,
+                data,
+            } => {
+                if !matches!(rpc_state.pending, Some(QueryHeader { id, .. }) if id == response.id) {
+                    openmina_core::error!(meta.time(); "pending query does not match the response");
+                    return Ok(());
+                }
+                let stream_id = rpc_state.stream_id;
+                let addr = rpc_state.addr;
+                let dispatcher = state_context.into_dispatcher();
+
+                dispatcher.push(P2pNetworkRpcAction::PrunePending {
+                    peer_id: *peer_id,
+                    stream_id,
+                });
+                dispatcher.push(P2pNetworkRpcAction::OutgoingData {
+                    addr,
+                    peer_id: *peer_id,
+                    stream_id,
+                    data: Data::from(
+                        RpcMessage::Response {
+                            header: response.clone(),
+                            bytes: data.clone(),
+                        }
+                        .into_bytes(),
+                    ),
+                    fin: false,
+                });
+                Ok(())
+            }
         }
     }
 
@@ -134,4 +364,207 @@ impl P2pNetworkRpcState {
             Ok(())
         }
     }
+}
+
+fn rpc_query_effects<'a, State, Action>(
+    peer_id: PeerId,
+    QueryHeader { tag, version, id }: &'a QueryHeader,
+    mut bytes: &[u8],
+    dispatcher: &mut Dispatcher<Action, State>,
+) -> Result<(), RpcQueryError<'a>>
+where
+    State: crate::P2pStateTrait,
+    Action: crate::P2pActionTrait<State>,
+{
+    let id = *id;
+    match (tag.as_ref(), *version) {
+        (rpc::GetBestTipV2::NAME, rpc::GetBestTipV2::VERSION) => {
+            rpc::GetBestTipV2::query_payload(&mut bytes)?;
+            dispatcher.push(P2pChannelsRpcAction::RequestReceived {
+                peer_id,
+                id,
+                request: Box::new(P2pRpcRequest::BestTipWithProof),
+            });
+        }
+        (rpc::AnswerSyncLedgerQueryV2::NAME, rpc::AnswerSyncLedgerQueryV2::VERSION) => {
+            let (hash, query) = rpc::AnswerSyncLedgerQueryV2::query_payload(&mut bytes)?;
+            let hash = v2::LedgerHash::from(v2::MinaBaseLedgerHash0StableV1(hash));
+
+            dispatcher.push(P2pChannelsRpcAction::RequestReceived {
+                peer_id,
+                id,
+                request: Box::new(P2pRpcRequest::LedgerQuery(hash, query)),
+            });
+        }
+        (
+            rpc::GetStagedLedgerAuxAndPendingCoinbasesAtHashV2::NAME,
+            rpc::GetStagedLedgerAuxAndPendingCoinbasesAtHashV2::VERSION,
+        ) => {
+            let hash =
+                rpc::GetStagedLedgerAuxAndPendingCoinbasesAtHashV2::query_payload(&mut bytes)?;
+            let hash = v2::StateHash::from(v2::DataHashLibStateHashStableV1(hash));
+            let request = Box::new(P2pRpcRequest::StagedLedgerAuxAndPendingCoinbasesAtBlock(
+                hash,
+            ));
+
+            dispatcher.push(P2pChannelsRpcAction::RequestReceived {
+                peer_id,
+                id,
+                request,
+            });
+        }
+        (rpc::GetTransitionChainV2::NAME, rpc::GetTransitionChainV2::VERSION) => {
+            let hashes = rpc::GetTransitionChainV2::query_payload(&mut bytes)?;
+            for hash in hashes {
+                let hash = v2::StateHash::from(v2::DataHashLibStateHashStableV1(hash));
+
+                dispatcher.push(P2pChannelsRpcAction::RequestReceived {
+                    peer_id,
+                    id,
+                    request: Box::new(P2pRpcRequest::Block(hash)),
+                });
+            }
+        }
+        (rpc::GetSomeInitialPeersV1ForV2::NAME, rpc::GetSomeInitialPeersV1ForV2::VERSION) => {
+            let () = rpc::GetSomeInitialPeersV1ForV2::query_payload(&mut bytes)?;
+            dispatcher.push(P2pChannelsRpcAction::RequestReceived {
+                peer_id,
+                id,
+                request: Box::new(P2pRpcRequest::InitialPeers),
+            });
+        }
+        (name, version) => return Err(RpcQueryError::Unimplemented(name, version)),
+    }
+    Ok(())
+}
+
+fn rpc_response_effects<State, Action>(
+    peer_id: PeerId,
+    QueryHeader { tag, version, id }: &QueryHeader,
+    mut bytes: &[u8],
+    dispatcher: &mut Dispatcher<Action, State>,
+) -> Result<(), RpcResponseError>
+where
+    State: crate::P2pStateTrait,
+    Action: crate::P2pActionTrait<State>,
+{
+    let id = *id;
+    match (tag.as_ref(), *version) {
+        (rpc::GetBestTipV2::NAME, rpc::GetBestTipV2::VERSION) => {
+            let response = rpc::GetBestTipV2::response_payload(&mut bytes)?
+                .map(|resp| BestTipWithProof {
+                    best_tip: resp.data.into(),
+                    proof: (resp.proof.0, resp.proof.1.into()),
+                })
+                .map(P2pRpcResponse::BestTipWithProof)
+                .map(Box::new);
+
+            dispatcher.push(P2pChannelsRpcAction::ResponseReceived {
+                peer_id,
+                id,
+                response,
+            });
+        }
+        (rpc::AnswerSyncLedgerQueryV2::NAME, rpc::AnswerSyncLedgerQueryV2::VERSION) => {
+            let response = Result::from(rpc::AnswerSyncLedgerQueryV2::response_payload(
+                &mut bytes,
+            )?)
+            .map_err(|e| RpcResponseError::Other {
+                rpc_id: rpc::AnswerSyncLedgerQueryV2::rpc_id(),
+                error: e.to_string(),
+            })?;
+            let response = Some(Box::new(P2pRpcResponse::LedgerQuery(response)));
+
+            dispatcher.push(P2pChannelsRpcAction::ResponseReceived {
+                peer_id,
+                id,
+                response,
+            });
+        }
+        (
+            rpc::GetStagedLedgerAuxAndPendingCoinbasesAtHashV2::NAME,
+            rpc::GetStagedLedgerAuxAndPendingCoinbasesAtHashV2::VERSION,
+        ) => {
+            let response =
+                rpc::GetStagedLedgerAuxAndPendingCoinbasesAtHashV2::response_payload(&mut bytes)?;
+            let response = response
+                .map(|(scan_state, hash, pending_coinbase, needed_blocks)| {
+                    let staged_ledger_hash = v2::MinaBaseLedgerHash0StableV1(hash).into();
+                    Arc::new(StagedLedgerAuxAndPendingCoinbases {
+                        scan_state,
+                        staged_ledger_hash,
+                        pending_coinbase,
+                        needed_blocks,
+                    })
+                })
+                .map(P2pRpcResponse::StagedLedgerAuxAndPendingCoinbasesAtBlock)
+                .map(Box::new);
+
+            dispatcher.push(P2pChannelsRpcAction::ResponseReceived {
+                peer_id,
+                id,
+                response,
+            });
+        }
+        (rpc::GetTransitionChainV2::NAME, rpc::GetTransitionChainV2::VERSION) => {
+            let response = rpc::GetTransitionChainV2::response_payload(&mut bytes)?;
+            match response {
+                Some(response) if !response.is_empty() => {
+                    for block in response {
+                        let response = Some(Box::new(P2pRpcResponse::Block(Arc::new(block))));
+                        dispatcher.push(P2pChannelsRpcAction::ResponseReceived {
+                            peer_id,
+                            id,
+                            response,
+                        });
+                    }
+                }
+                _ => {
+                    dispatcher.push(P2pChannelsRpcAction::ResponseReceived {
+                        peer_id,
+                        id,
+                        response: None,
+                    });
+                }
+            }
+        }
+        (rpc::GetSomeInitialPeersV1ForV2::NAME, rpc::GetSomeInitialPeersV1ForV2::VERSION) => {
+            let response = rpc::GetSomeInitialPeersV1ForV2::response_payload(&mut bytes)?;
+            if response.is_empty() {
+                dispatcher.push(P2pChannelsRpcAction::ResponseReceived {
+                    peer_id,
+                    id,
+                    response: None,
+                });
+            } else {
+                let peers = response
+                    .into_iter()
+                    .filter_map(P2pConnectionOutgoingInitOpts::try_from_mina_rpc)
+                    .collect();
+                dispatcher.push(P2pChannelsRpcAction::ResponseReceived {
+                    peer_id,
+                    id,
+                    response: Some(Box::new(P2pRpcResponse::InitialPeers(peers))),
+                });
+            }
+        }
+        _ => {}
+    }
+    Ok(())
+}
+
+#[derive(Debug, thiserror::Error)]
+enum RpcQueryError<'a> {
+    #[error(transparent)]
+    Read(#[from] RpcQueryReadError),
+    #[error("unimplemented rpc {}:{1}", String::from_utf8_lossy(.0))]
+    Unimplemented(&'a [u8], Ver),
+}
+
+#[derive(Debug, thiserror::Error)]
+enum RpcResponseError {
+    #[error(transparent)]
+    Read(#[from] RpcResponseReadError),
+    #[error("rpc response {rpc_id} error: {error}")]
+    Other { rpc_id: String, error: String },
 }
